Revert "feat: Generate full API url for uploaded file on the server (#1275)"
This reverts commit 74cc7e30c6e0c64bd23c0fa9ad26a15cb3bfb949.
