+++ api.planx.uk/s3/uploadFile.ts
@@ -13,11 +13,10 @@ export const uploadPublicFile = async (
   const { params, key, fileType } = generateFileParams(file, filename);
 
   await s3.putObject(params).promise();
-  const fileUrl = buildFileUrl(key, "public")
 
   return {
     file_type: fileType,
-    fileUrl,
+    key,
   };
 };
 
@@ -34,21 +33,11 @@ export const uploadPrivateFile = async (
   };
 
   await s3.putObject(params).promise();
-  const fileUrl = buildFileUrl(key, "private")
 
   return {
     file_type: fileType,
-    fileUrl,
+    key,
   };
-};
-
-// Construct an API URL for the uploaded file
-const buildFileUrl = (key: string, path: "public" | "private") => {
-  const s3 = s3Factory();
-  const s3Url = new URL(s3.getSignedUrl('getObject', { Key: key }));
-  // URL.pathname has a leading "/"
-  const fileUrl = `${process.env.API_URL_EXT}/file/${path}${s3Url.pathname}`
-  return fileUrl;
 };
 
 export function generateFileParams(
